Change function for loading explicit area of burn to work...
... with 64bit integers. If we store AOB as m-2, then we need
64bit integers to accomadoate the values.

+++ include/WildFire.h
@@ -97,7 +97,7 @@ private:
   std::vector<int> exp_burn_mask;
   std::vector<int> exp_jday_of_burn;
   std::vector<int> exp_fire_severity;
-  std::vector<int> exp_area_of_burn;
+  std::vector<int64_t> exp_area_of_burn;
